refactor: move around file types, track types
