+++ midix/src/file/mod.rs
@@ -1,1 +1,3 @@
-
+pub mod chunk;
+pub mod header;
+pub mod track;
