+++ midix/src/lib.rs
@@ -32,8 +32,8 @@ mod error;
 
 pub mod bytes;
 pub mod channel;
-pub mod chunk;
 pub mod file;
+pub mod live;
 pub mod message;
 pub mod reader;
 pub(crate) mod utils;
@@ -41,6 +41,8 @@ pub(crate) mod utils;
 pub mod prelude {
     pub use crate::bytes::*;
     pub use crate::channel::Channel;
+    pub use crate::file::{chunk::*, header::*, track::*};
+    pub use crate::live::*;
     pub use crate::message::{controller::*, key::*, pitch_bend::*, program::*, velocity::*, *};
     pub use crate::reader::{error::*, *};
