+++ midix/src/message/mod.rs
@@ -3,17 +3,13 @@ pub mod key;
 pub mod pitch_bend;
 pub mod program;
 pub mod velocity;
-mod voice_message;
-use std::io::ErrorKind;
 
+mod voice_message;
 pub use voice_message::*;
+
 mod common_message;
 pub use common_message::*;
 mod realtime_message;
 pub use realtime_message::*;
-mod live;
-pub use live::*;
-mod track;
-pub use track::*;
 mod sysex;
 pub use sysex::*;
