--- midix/src/message/track.rs
@@ -1,8 +1,0 @@
-use crate::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub enum TrackMidiMessage {
-    ChannelVoice(ChannelVoiceMessage),
-    SystemExclusive(Vec<u8>),
-    SystemRealTime(SystemRealTimeMessage),
-}
